Added API Routes (with authentication for later login component). Removed GraphQL for now, as I don't think it's a good fit for storing large neural network models. Set up some blank routes in client's App.js. Created User model with queries (login and getUser) and create (register).

+++ server/models/Neuron.js
@@ -1,7 +1,7 @@
 const { Schema, model } = require('mongoose');
 
 const neuronSchema = new Schema({
-  weights: [Float],
+  weights: [Decimal128],
 });
 
 const Neuron = model('Neuron', neuronSchema);
